requirements.txt
arabic-reshaper==2.1.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
charset-normalizer==2.0.3
cssselect2==0.4.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.2
django-autoslug==1.9.8
django-cors-headers==3.10.1
django-crum==0.7.5
django-extra-fields==3.0.2
django-import-export==2.8.0
django-rest-passwordreset==1.2.1
django-widget-tweaks==1.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-extra-fields==3.2.1
et-xmlfile==1.0.1
future==0.18.2
html5lib==1.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
mssql-django==1.1.3
odfpy==1.4.1
openpyxl==3.0.6
pdfkit==0.6.1
Pillow==9.0.1
pycparser==2.20
PyJWT==2.3.0
pyodbc==4.0.32
Pyphen==0.10.0
python-bidi==0.4.2
python-decouple==3.6
pytz==2019.3
PyYAML==6.0
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
tablib==3.2.1
tinycss2==1.1.0
tzdata==2022.1
urllib3==1.26.2
WeasyPrint==52.2
webencodings==0.5.1
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
